requirements.txt
opencv-python
lxml
tqdm
tensorflow==2.3.0rc0
absl-py
easydict
matplotlib
pillow
